Update to enable HTML map export and add AsteroidBelt class
The commit enables the export of the abyssal map as an HTML file. An AsteroidBelt class has also been added to the Map_Components.py file. This class generates the density based on an asteroid belt's orbit and adds density to the asteroid belt's properties. Additionally, a minor text edit was made to correct a comment in the mass and density calculation section.