Add gravity attribute to Planet class
Included a gravity attribute to the Planet class in Map_Components.py. This addition required creating a method to generate the gravity based on the planet's mass and radius, and adding the attribute to the string representation and serialization of the planet object. Also made minor adjustments to print statements in other classes for cleaner output.